dbhydroR

dbhydroR is an R interface to the South Florida Water Management District's DBHYDRO database which holds over 35 million hydrologic and water quality records from the Florida Everglades and surrounding areas.

Hydrologic Data

The get_hydro function provides the capability to return:

weather data

surfacewater data

groundwater data

water-quality sonde data

Water Quality Data

The get_wq function provides the capability to return:

water quality data