Query dbkey information

Retrieve a data.frame summary including dbkeys or a vector of dbkeys corresponding to specified parameters

get_dbkey(category, stationid = NA, param = NA, freq = NA,
  longest = FALSE, stat = NA, recorder = NA, agency = NA,
  strata = NA, detail.level = "summary", ...)

Arguments

category
character string, choice of "WEATHER", "SW", "GW", or "WQ"

stationid
character string specifying station name

param
character string specifying desired parameter name

freq
character string specifying collection frequency (daily = "DA")

longest
logical limit results to the longest period-of-record?

stat
character string specifying statistic type

recorder
character string specifying recorder information

agency
character string specifying collector agency

strata
numeric vector of length 2 specifying a range of z-coordinates relative to local ground elevation. Only applicable for queries in the "WEATHER" and "GW" categories.

detail.level
character string specifying the level of detail to return. Choices are "full", "summary", and "dbkey".

...
Options passed as named parameters

Details

A dbkey represents a unique station x variable time-series. A value in the "Recorder" field of "PREF" should be used whenever possible. This indicates that the dataset has been checked by the SFWMD modelling group.

References

http://my.sfwmd.gov/dbhydroplsql/show_dbkey_info.main_menu

http://my.sfwmd.gov/dbhydroplsql/show_dbkey_info.show_meta_data

Examples

# NOT RUN {
# Weather
get_dbkey(stationid = "JBTS", category = "WEATHER", param = "WNDS",
detail.level = "summary")
get_dbkey(stationid = "JBTS", category = "WEATHER", param = "WNDS",
detail.level = "dbkey")

# query on multiple values
get_dbkey(stationid = c("MBTS", "JBTS"), category = "WEATHER",
param = "WNDS", freq = "DA", detail.level = "dbkey")

# Surfacewater
get_dbkey(stationid = "C111%", category = "SW")
get_dbkey(category = "SW", stationid = "LAKE%", detail.level = "full")

# Groundwater
get_dbkey(stationid = "C111%", category = "GW")
get_dbkey(stationid = "C111AE", category = "GW", param = "WELL",
freq = "DA", stat = "MEAN", strata = c(9, 22), recorder = "TROL",
 agency = "WMD", detail.level = "full")

# Water Quality
get_dbkey(stationid = "C111%", category = "WQ")
# }