Retrieve hydrologic data from the DBHYDRO Environmental Database

Retrieve hydrologic data from the DBHYDRO Environmental Database

get_hydro(dbkey = NA, date_min = NA, date_max = NA, raw = FALSE, ...)

Arguments

dbkey
character string specifying a unique data series. See get_dbkey

date_min
character date must be in YYYY-MM-DD format

date_max
character date must be in YYYY-MM-DD format

raw
logical default is FALSE, set to TRUE to return data in "long" format with all comments, qa information, and database codes included.

...
Options passed on to get_dbkey

Details

get_hydro can be run in one of two ways.

The first, is to identify one or more dbkeys before-hand that correspond to unique data series and are passed to the dbkey argument. dbkeys can be found by:

iterative calls to get_dbkey (see example)

using the Environmental Monitoring Location Maps (https://www.sfwmd.gov/documents-by-tag/emmaps)

using the DBHYDRO Browser (http://my.sfwmd.gov/dbhydroplsql/show_dbkey_info.main_menu).

The second way to run get_hydro is to specify additional arguments to ... which are passed to get_dbkey on-the-fly.

By default, get_hydro returns a cleaned output where metadata (station-name, variable, measurement units) is wholly contained in the column name. This is accomplished internally by the clean_hydro function. If additional metadata such as latitude and longitude are desired set the raw argument to TRUE.

Examples

# NOT RUN {
#One variable/station time series
get_hydro(dbkey = "15081", date_min = "2013-01-01", date_max = "2013-02-02")

#Multiple variable/station time series
get_hydro(dbkey = c("15081", "15069"),
date_min = "2013-01-01", date_max = "2013-02-02")

#Instantaneous hydro retrieval
get_hydro(dbkey = "IY639", date_min = "2015-11-01", date_max = "2015-11-04")

#Looking up unknown dbkeys on the fly
get_hydro(stationid = "JBTS", category = "WEATHER",
param = "WNDS", freq = "DA", date_min = "2013-01-01",
date_max = "2013-02-02")
# }