Retrieve water quality data from the DBHYDRO Environmental Database

Retrieve water quality data from the DBHYDRO Environmental Database

get_wq(station_id = NA, date_min = NA, date_max = NA, test_name = NA,
  mdl_handling = "raw", raw = FALSE, qc_strip = "N", qc_field = "N",
  test_number = NA, v_target_code = "file_csv", sample_id = NA,
  project_code = NA)

Arguments

station_id
character string of station id(s). See the SFWMD station search utility at http://my.sfwmd.gov/dbhydroplsql/water_quality_data.show_group_station_characters for specific options

date_min
character date must be in POSIXct YYYY-MM-DD format

date_max
character date must be in POSIXct YYYY-MM-DD format

test_name
character string of test name(s). See the ArcGIS Online Station Map at http://my.sfwmd.gov/WAB/EnvironmentalMonitoring/index.html for specific options

mdl_handling
character string specifying the handling of measurement values below the minimum detection limit (MDL). Example choices for this argument include:

raw: Returns values exactly as they are stored in the database. Current practice is to return values below the MDL as 0 minus the uncertainty estimate.

half: Returns values below the MDL as half the MDL

full: Returns values below the MDL as the MDL

raw
logical default is FALSE, set to TRUE to return data in "long" format with all comments, qa information, and database codes included

qc_strip
logical set TRUE to avoid returning QAQC flagged data entries

qc_field
logical set TRUE to avoid returning field QC results

test_number
numeric test name alternative (not implemented)

v_target_code
string print to file? (not implemented)

sample_id
numeric (not implemented)

project_code
numeric (not implemented)

Details

By default, get_wq returns a cleaned output. First, the cleaning function clean_wq converts the raw output from native DBHYDRO long format (each piece of data on its own row) to wide format (each site x variable combination in its own column). Next, the extra columns associated with QA flags, LIMS, and District receiving are removed. Finally, row entries associated with QA field blanks, which are used to check on potential sources of contamination, are removed. Setting the raw flag to TRUE will force getwq to retain information on QA field blanks as well as the other QA fields.

Examples

#one variable and one station
get_wq(station_id = "FLAB08",
date_min = "2011-03-01", date_max = "2012-05-01",
test_name = "CHLOROPHYLLA-SALINE")

#>          date FLAB08_CHLOROPHYLLA-SALINE_ug/L
#> 1  2011-03-21                           0.196
#> 2  2011-04-11                           0.276
#> 3  2011-05-09                           0.328
#> 4  2011-06-07                           0.329
#> 5  2011-07-05                           0.402
#> 6  2011-08-08                           0.907
#> 7  2011-09-06                           0.887
#> 8  2011-10-06                           0.573
#> 9  2011-12-08                           0.369
#> 10 2012-02-02                           0.260
#> 11 2012-04-05                           0.509

# NOT RUN {
#one variable at multiple stations
get_wq(station_id = c("FLAB08", "FLAB09"),
date_min = "2011-03-01", date_max = "2012-05-01",
test_name = "CHLOROPHYLLA-SALINE")

#One variable at a wildcard station
get_wq(station_id = c("FLAB0%"),
date_min = "2011-03-01",
date_max = "2012-05-01",
test_name = "CHLOROPHYLLA-SALINE")

#multiple variables at multiple stations
get_wq(station_id = c("FLAB08", "FLAB09"),
date_min = "2011-03-01", date_max = "2012-05-01",
test_name = c("CHLOROPHYLLA-SALINE", "SALINITY"))
# }